--- a/src/hotspot/share/runtime/deoptimization.cpp
+++ b/src/hotspot/share/runtime/deoptimization.cpp
@@ -296,32 +296,31 @@ static void restore_eliminated_locks(JavaThread* thread, GrowableArray<compiledV
                                                      exec_mode, realloc_failures);
       deoptimized_objects = deoptimized_objects || relocked;
 #ifndef PRODUCT
-      if (PrintDeoptimizationDetails) {
-        ResourceMark rm;
-        stringStream st;
+      LogMessage(deoptimization) lm;
+      if (lm.is_debug()) {
+        NonInterleavingLogStream ls(LogLevelType::Debug, lm);
         for (int j = 0; j < monitors->length(); j++) {
           MonitorInfo* mi = monitors->at(j);
           if (mi->eliminated()) {
             if (first) {
               first = false;
-              st.print_cr("RELOCK OBJECTS in thread " INTPTR_FORMAT, p2i(thread));
+              ls.print_cr("RELOCK OBJECTS in thread " INTPTR_FORMAT, p2i(thread));
             }
             if (exec_mode == Deoptimization::Unpack_none) {
               ObjectMonitor* monitor = deoptee_thread->current_waiting_monitor();
               if (monitor != NULL && monitor->object() == mi->owner()) {
-                st.print_cr("     object <" INTPTR_FORMAT "> DEFERRED relocking after wait", p2i(mi->owner()));
+                ls.print_cr("     object <" INTPTR_FORMAT "> DEFERRED relocking after wait", p2i(mi->owner()));
                 continue;
               }
             }
             if (mi->owner_is_scalar_replaced()) {
               Klass* k = java_lang_Class::as_Klass(mi->owner_klass());
-              st.print_cr("     failed reallocation for klass %s", k->external_name());
+              ls.print_cr("     failed reallocation for klass %s", k->external_name());
             } else {
-              st.print_cr("     object <" INTPTR_FORMAT "> locked", p2i(mi->owner()));
+              ls.print_cr("     object <" INTPTR_FORMAT "> locked", p2i(mi->owner()));
             }
           }
         }
-        tty->print_raw(st.freeze());
       }
 #endif // !PRODUCT
     }
@@ -439,9 +438,7 @@ Deoptimization::UnrollBlock* Deoptimization::fetch_unroll_info_helper(JavaThread
   Handle exceptionObject;
   if (trap_scope->rethrow_exception()) {
 #ifndef PRODUCT
-    if (PrintDeoptimizationDetails) {
-      tty->print_cr("Exception to be rethrown in the interpreter for method %s::%s at bci %d", trap_scope->method()->method_holder()->name()->as_C_string(), trap_scope->method()->name()->as_C_string(), trap_scope->bci());
-    }
+    log_debug(deoptimization)("Exception to be rethrown in the interpreter for method %s::%s at bci %d", trap_scope->method()->method_holder()->name()->as_C_string(), trap_scope->method()->name()->as_C_string(), trap_scope->bci());
 #endif // !PRODUCT
 
     GrowableArray<ScopeValue*>* expressions = trap_scope->expressions();
@@ -1449,9 +1446,7 @@ void Deoptimization::reassign_fields(frame* fr, RegisterMap* reg_map, GrowableAr
     Handle obj = sv->value();
     assert(obj.not_null() || realloc_failures, "reallocation was missed");
 #ifndef PRODUCT
-    if (PrintDeoptimizationDetails) {
-      tty->print_cr("reassign fields for object of type %s!", k->name()->as_C_string());
-    }
+    log_debug(deoptimization)("reassign fields for object of type %s!", k->name()->as_C_string());
 #endif // !PRODUCT
 
     if (obj.is_null()) {
@@ -1471,11 +1466,13 @@ void Deoptimization::reassign_fields(frame* fr, RegisterMap* reg_map, GrowableAr
       if (payload->is_location() &&
           payload->as_LocationValue()->location().type() == Location::vector) {
 #ifndef PRODUCT
-        if (PrintDeoptimizationDetails) {
-          tty->print_cr("skip field reassignment for this vector - it should be assigned already");
-          if (Verbose) {
+        log_debug(deoptimization)("skip field reassignment for this vector - it should be assigned already");
+        {
+          LogMessage(deoptimization) lm;
+          if (lm.is_debug()) {
+            NonInterleavingLogStream ls(LogLevelType::Debug, lm);
             Handle obj = sv->value();
-            k->oop_print_on(obj(), tty);
+            k->oop_print_on(obj(), &ls);
           }
         }
 #endif // !PRODUCT