Hi!

This PR ports PrintDeoptimizationDetails to UL by mapping its output to debug level with tag deoptimization.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8295060](https://bugs.openjdk.org/browse/JDK-8295060): Port PrintDeoptimizationDetails to UL


### Reviewers
 * [Coleen Phillimore](https://openjdk.org/census#coleenp) (@coleenp - **Reviewer**) ⚠️ Review applies to [5b9f53d3](https://git.openjdk.org/jdk/pull/10645/files/5b9f53d33291c3c23ccbc8aceb1662a0d25bc0ea)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk pull/10645/head:pull/10645` \
`$ git checkout pull/10645`

Update a local copy of the PR: \
`$ git checkout pull/10645` \
`$ git pull https://git.openjdk.org/jdk pull/10645/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 10645`

View PR using the GUI difftool: \
`$ git pr show -t 10645`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/10645.diff">https://git.openjdk.org/jdk/pull/10645.diff</a>

</details>