@jdksjolen 2022-10-19
> Nice work. Can you show an example of new output vs old?

New: `-Xlog:deoptimization=trace:deopt.txt`

```
[1,923s][debug][deoptimization] Expressions size: 1
[1,923s][debug][deoptimization]  - Reconstructed expression 0 (OBJECT): NULL
[1,923s][debug][deoptimization] Locals size: 3
[1,923s][debug][deoptimization]  - Reconstructed local 0 (OBJECT): java/lang/String
[1,923s][debug][deoptimization]  - Reconstructed local 1 (OBJECT): NULL
[1,924s][debug][deoptimization] [1. Interpreted Frame]
[1,924s][debug][deoptimization] Interpreted frame (sp=0x00007fd07e205548 unextended sp=0x00007fd07e205548, fp=0x00007fd07e205598, real_fp=0x00007fd07e205598, pc=0x00007fd0704506a0)
[1,924s][debug][deoptimization] ~deoptimization entry points  [0x00007fd0704506a0, 0x00007fd070453b38]  13464 bytes
[1,924s][debug][deoptimization]      BufferBlob (0x00007fd070431510) used for Interpreter
[1,924s][debug][deoptimization]  - local  [0x000000011d00d0e8]; #0
[1,924s][debug][deoptimization]  - local  [0x0000000000000000]; #1
[1,924s][debug][deoptimization]  - local  [0x0000000000000000]; #2
[1,924s][debug][deoptimization]  - stack  [0x0000000000000000]; #0
[1,924s][debug][deoptimization]  - monitor[0x00007fd07e205550]
[1,924s][debug][deoptimization]  - bcp    [0x00007fd02800b230]; @8
[1,924s][debug][deoptimization]  - locals [0x00007fd07e2055b8]
[1,924s][debug][deoptimization]  - method [0x00007fd02800b298]; virtual jboolean java.lang.String.equals(jobject)
[1,924s][debug][deoptimization] 
```

Old: 

``` `-XX:+WizardMode -XX:+Verbose -XX:+PrintDeoptimizationDetails`
Expressions size: 1
 - Reconstructed expression 0 (OBJECT): NULL
Locals size: 3
 - Reconstructed local 0 (OBJECT): java/lang/String
 - Reconstructed local 1 (OBJECT): NULL
[1. Interpreted Frame]
Interpreted frame (sp=0x00007fc830059548 unextended sp=0x00007fc830059548, fp=0x00007fc830059598, real_fp=0x00007fc830059598, pc=0x00007fc8204506a0)
~deoptimization entry points  [0x00007fc8204506a0, 0x00007fc820453b38]  13464 bytes
     BufferBlob (0x00007fc820431510) used for Interpreter
Trying to load: /home/johan/jdk/build/linux-x64-slowdebug/jdk/lib/server/libhsdis-amd64.so
Trying to load: /home/johan/jdk/build/linux-x64-slowdebug/jdk/lib/server/hsdis-amd64.so
Trying to load: /home/johan/jdk/build/linux-x64-slowdebug/jdk/lib/hsdis-amd64.so
Trying to load: hsdis-amd64.so
[1,857s][warning][os] Loading hsdis library failed
Could not load hsdis-amd64.so; hsdis-amd64.so: cannot open shared object file: No such file or directory; PrintAssembly defaults to abstract disassembly.
[MachCode]
// Snip! Huge amount of logs here
[/MachCode]
```