@vnkozlov 2022-10-19
> Old `-XX:+WizardMode -XX:+Verbose -XX:+PrintDeoptimizationDetails`:

You only need `-XX:+PrintDeoptimizationDetails` to see old output. `-XX:+WizardMode -XX:+Verbose` will mess it up.

I don't see next output in your example after `- method ` last line:
```
        {method} {0x0000000800009b10} 'equals' '(Ljava/lang/Object;)Z' in 'java/lang/String'
        bci:    8
        locals:
           0 "true"{0x00000007ff84c3a0} <0x00000007ff84c3a0>
           1 NULL <0x0000000000000000>
           2 0 (int) 0.000000 (float) 0 (hex)
        expressions:
           0 NULL <0x0000000000000000>
```