@vnkozlov 2022-10-19
I see in new output next lines repeated instead of missing lines from old output:
```
0.703s][debug][deoptimization] Interpreted frame (sp=0x00007fb5bbf165b8 unextended sp=0x00007fb5bbf165b8, fp=0x00007fb5bbf16608, real_fp=0x00007fb5bbf16608, pc=0x00007fb5a43c26a0)
[0.703s][debug][deoptimization] ~deoptimization entry points  [0x00007fb5a43c26a0, 0x00007fb5a43c5b38]  13464 bytes
[0.703s][debug][deoptimization]      BufferBlob (0x00007fb5a43a3510) used for Interpreter
```