@jdksjolen 2022-10-27
I made a new comparison and found this:

- UL adds a line, which comes from this PR: https://github.com/openjdk/jdk/pull/8812
- UL has the precise description of locals missing, looks like this in PrintDeoptimizationDetails:

```
	locals:
	   0 "true"{0x000000011d00cd40} <0x000000011d00cd40>
	   1 NULL <0x0000000000000000>
	   2 0 (int) 0,000000 (float) 0 (hex)
	expressions:
	   0 NULL <0x0000000000000000>
```

No clue why this is missing so far, looking into it.