@jdksjolen 2022-11-08
Regarding:

>UL has the precise description of locals missing, looks like this in PrintDeoptimizationDetails:

This is not printed because that info comes from `javaVFrame::print`. `vframe::new_vframe` can return quite a few different `vframe` subclasses, and as such each of these needs to have their `print` be refactored into `print_on + print`.